+++ src/ui/src/js/controllers/request_index.js
@@ -94,7 +94,7 @@ export default function requestIndexController(
       .withOption('newData', true)
       .withPaginationType('full_numbers')
       .withBootstrap()
-      .withOption('searchDelay', 1200)
+      .withOption('searchDelay', $scope.config.searchDelay)
       .withOption('createdRow', function(row, data, dataIndex) {
         $compile(angular.element(row).contents())($scope);
       });
@@ -177,43 +177,61 @@ export default function requestIndexController(
     $('#requestIndexTable').on('length.dt', (event, settings, len) => {
       localStorageService.set('_request_index_length', len);
     });
+
+    let filtering = null;
+    $('.dataTables_filter input[type=search]')
+      .off('keyup.DT search.DT input.DT paste.DT cut.DT')
+      .on('input.DT search.DT paste.DT cut.DT', function(e) {
+        if (filtering) {
+          clearTimeout(filtering);
+        }
+        filtering = setTimeout(function() {
+          let searchVal = $('.dataTables_filter input[type=search]').val()
+          if (searchVal.length > 2) {
+            $.fn['dataTable'].tables({ api: true }).search(searchVal).draw();
+          }
+          if (searchVal.length == 0) {
+            $.fn['dataTable'].tables({ api: true }).search(searchVal).draw();
+          }
+        }, $scope.config.searchDelay);
+      });
   };
 
   const lightColumnFilterOptions = {
     command_display_name: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'Command Filter'},
     },
     namespace: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'Namespace Filter'},
     },
     system: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'System Filter'},
     },
     system_version: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'Version Filter'},
     },
     instance_name: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'Instance Filter'},
     },
     requester: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'Requester Filter'},
     },
     status: {
@@ -255,7 +273,7 @@ export default function requestIndexController(
     comment: {
       html: 'input',
       type: 'text',
-      time: 600,
+      time: $scope.config.searchDelay,
       attr: {class: 'form-inline form-control', title: 'Comment Filter'},
     },
     metadata: {},