+++ CHANGELOG.md
@@ -68,8 +68,10 @@ TBD
 - Updated Admin Garden Page to load Garden models from preloaded Garden models
 - Fixed File validate to query for chunks and ensure accurate size
 - Fixed bug where API updates to Instance Status did not publish topic request
+- Request search delay configurable through `ui.search_delay` config option.
 - Removed status, status_info, and namespaces from Garden
 - Update Request Pruning to utilize new field Expiration At instead of each command type independently 
+- Update Requests index search box from text_search to or filter on UI columns
 
 # 3.29.1